# Optimize moe_sum_reduce_kernel