<!-- Thank you for your contribution! Please follow these guidelines to enhance your pull request. If anything is unclear, submit your PR and reach out to maintainers for assistance. Join our Slack community at https://slack.sglang.ai to discuss further. -->

## Motivation

During the Prefill stage of TP MoE, the moe_sum_reduce_kernel operator accounts for a relatively large proportion.
<img width="2996" height="1606" alt="image" src="https://github.com/user-attachments/assets/c49f154e-92dc-41a5-a0e5-dbf5fc645b87" />

This PR is to optimize moe_sum_reduce_kernel and get speedup up to 17.6%.
DeepSeek-V3 E2E throughput speedup 5-10%.

============Main============
TritonKernel column is the concerned kernel.
```
Running correctness verification...
✅ All implementations match

Running performance benchmark...
sum_scaled_performance:
    num_tokens    Original  TorchCompile  TritonKernel
0          1.0    9.920000      5.936000      7.840000
1          2.0   10.368000      6.080000      8.032000
2          4.0   10.560000      6.432000      7.904000
3          8.0   10.752000      7.072000      8.128000
4         16.0   11.008000      8.704000      8.608000
5         32.0   11.360000     12.000000      8.960000
6         64.0   12.544000     18.015999      9.440000
7        128.0   15.807999     30.432001     10.688000
8        256.0   21.407999     55.328000     14.048000
9        512.0   34.688000    104.128003     21.888001
10      1024.0   58.591999    202.575997     35.840001
11      2048.0  101.712000    398.575991     59.519999
12      4096.0  189.536005    792.576015    104.032002
```

============This PR============
```
Running correctness verification...
✅ All implementations match

Running performance benchmark...
sum_scaled_performance:
    num_tokens    Original  TorchCompile  TritonKernel
0          1.0    9.536000      5.824000      6.688000
1          2.0    9.984000      5.856000      6.816000
2          4.0   10.144000      6.336000      6.784000
3          8.0   10.336000      6.976000      7.008000
4         16.0   10.592000      8.576000      7.328000
5         32.0   10.976000     11.616000      7.680000
6         64.0   12.128000     17.824000      8.480000
7        128.0   15.456000     30.336000     10.016000
8        256.0   21.088000     54.848000     14.304000
9        512.0   34.336001    104.128003     22.752000
10      1024.0   58.240000    202.207997     36.959998
11      2048.0  102.527998    400.736004     61.983999
12      4096.0  189.152002    791.040003    109.664001
```
<!-- Describe the purpose and goals of this pull request. -->

## Modifications

<!-- Detail the changes made in this pull request. -->

## Accuracy Tests

<!-- If this pull request affects model outputs (e.g., changes to the kernel or model forward code), provide accuracy test results. -->

## Benchmarking and Profiling
```
$python3 -m sglang.launch_server --model /home/admin/DeepSeek-V3 --tp-size 8 --port 30000 --mem-fraction-static 0.9

This PR:

First round warmup, is not high,
$python3 benchmark/gsm8k/bench_sglang.py --num-questions 200 --parallel 128 --num-shots 8 --port 30000 --data-path ./test.jsonl

100%|████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 200/200 [01:06<00:00,  3.00it/s]
Accuracy: 0.970
Invalid: 0.000
Latency: 68.177 s
Output throughput: 300.540 token/s

Second round on-ward, throughput increase 10%.
[root  /root/luoyuan.luo/workspace/SGLang] 日 8月 24 14:03:44 
$python3 benchmark/gsm8k/bench_sglang.py --num-questions 200 --parallel 128 --num-shots 8 --port 30000 --data-path ./test.jsonl
100%|████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 200/200 [00:17<00:00, 11.52it/s]
Accuracy: 0.970
Invalid: 0.000
Latency: 17.525 s
Output throughput: 1166.097 token/s

Main:
$python3 benchmark/gsm8k/bench_sglang.py --num-questions 200 --parallel 128 --num-shots 8 --port 30000 --data-path ./test.jsonl
100%|████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 200/200 [00:19<00:00, 10.29it/s]
Accuracy: 0.965
Invalid: 0.000
Latency: 19.708 s
Output throughput: 1033.465 token/s
```

## Checklist

- [ ] Format your code according to the [Format code with pre-commit](https://docs.sglang.ai/developer_guide/contribution_guide.html#format-code-with-pre-commit).
- [ ] Add unit tests according to the [Run and add unit tests](https://docs.sglang.ai/developer_guide/contribution_guide.html#run-and-add-unit-tests).
- [ ] Update documentation according to [Write documentations](https://docs.sglang.ai/developer_guide/contribution_guide.html#write-documentations).
- [ ] Provide accuracy and speed benchmark results according to [Test the accuracy](https://docs.sglang.ai/developer_guide/contribution_guide.html#test-the-accuracy) and [Benchmark the speed](https://docs.sglang.ai/developer_guide/contribution_guide.html#benchmark-the-speed).