@BBuf 2025-08-22
Please report DeepSeek-V3 gsm8k acc too.