@yuan-luo 2025-09-05
Encountered UT error in Xeon, WIP.
```
/sglang-checkout/python/sglang/srt/layers/quantization/awq.py:59: UserWarning: Only CUDA and HIP support AWQ currently.
  warnings.warn(f"Only CUDA and HIP support AWQ currently.")
Traceback (most recent call last):
  File "/sglang-checkout/test/srt/cpu/test_topk.py", line 13, in <module>
    from sglang.srt.models.llama4 import Llama4MoE
  File "/sglang-checkout/python/sglang/srt/models/llama4.py", line 42, in <module>
    from sglang.srt.layers.moe.fused_moe_triton import FusedMoE
  File "/sglang-checkout/python/sglang/srt/layers/moe/fused_moe_triton/__init__.py", line 4, in <module>
    from sglang.srt.layers.moe.fused_moe_triton.fused_moe import fused_experts
  File "/sglang-checkout/python/sglang/srt/layers/moe/fused_moe_triton/fused_moe.py", line 26, in <module>
    from .fused_moe_triton_kernels import invoke_fused_moe_kernel, moe_sum_reduce_triton
  File "/sglang-checkout/python/sglang/srt/layers/moe/fused_moe_triton/fused_moe_triton_kernels.py", line 807, in <module>
    _moe_sum_reduce_kernel = triton.autotune(
                             ^^^^^^^^^^^^^^^^
  File "/sgl-workspace/miniforge3/lib/python3.12/site-packages/triton/runtime/autotuner.py", line 378, in decorator
    return Autotuner(fn, fn.arg_names, configs, key, reset_to_zero, restore_value, pre_hook=pre_hook,
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/sgl-workspace/miniforge3/lib/python3.12/site-packages/triton/runtime/autotuner.py", line 130, in __init__
    self.do_bench = driver.active.get_benchmarker()
                    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/sgl-workspace/miniforge3/lib/python3.12/site-packages/triton/runtime/driver.py", line 23, in __getattr__
    self._initialize_obj()
  File "/sgl-workspace/miniforge3/lib/python3.12/site-packages/triton/runtime/driver.py", line 20, in _initialize_obj
    self._obj = self._init_fn()
                ^^^^^^^^^^^^^^^
  File "/sgl-workspace/miniforge3/lib/python3.12/site-packages/triton/runtime/driver.py", line 8, in _create_driver
    raise RuntimeError(f"{len(actives)} active drivers ({actives}). There should only be one.")
RuntimeError: 0 active drivers ([]). There should only be one.
```